names of the content extractor classes changed
ContentExtractor -> InternalContentExtractor
ContentExtractorWithTimeout -> ContentExtractor
