+++ cermine-impl/src/main/java/pl/edu/icm/cermine/PdfRawTextWithLabelsExtractor.java
@@ -107,7 +107,7 @@ public class PdfRawTextWithLabelsExtractor {
         
         String path = parser.getPath();
         String extension = parser.getTextExtension();
-        ContentExtractor.THREADS_NUMBER = parser.getThreadsNumber();
+        InternalContentExtractor.THREADS_NUMBER = parser.getThreadsNumber();
  
         File file = new File(path);
         if (file.isFile()) {
